+++ geotrek/tourism/admin.py
@@ -87,3 +87,10 @@ if settings.TOURISM_ENABLED:
     admin.site.register(tourism_models.TouristicEventType, TouristicEventTypeAdmin)
     admin.site.register(tourism_models.TouristicEventParticipantCategory, TouristicEventParticipantCategoryAdmin)
     admin.site.register(tourism_models.CancellationReason, CancellationReasonAdmin)
+
+class TouristicEventPlaceAdmin(LeafletGeoAdmin):
+    list_display = ('name',)
+    list_filter = ('name',)
+    search_fields = ('name',)
+
+admin.site.register(tourism_models.TouristicEventPlace, TouristicEventPlaceAdmin)