+++ geotrek/tourism/forms.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 
 from django.utils.translation import gettext_lazy as _
+from geotrek.tourism.widgets import AutoLocateMapWidget
+from mapentity.widgets import MapWidget
 
 from crispy_forms.layout import Div, HTML, Fieldset
 
@@ -84,6 +86,7 @@ class TouristicEventForm(CommonForm):
                     'meeting_point',
                     'start_time',
                     'end_time',
+                    'place',
                     'contact',
                     'email',
                     'website',
@@ -122,13 +125,14 @@ class TouristicEventForm(CommonForm):
     ]
 
     class Meta:
-        fields = ['name', 'review', 'published', 'description_teaser', 'description',
+        fields = ['name', 'place', 'review', 'published', 'description_teaser', 'description',
                   'themes', 'begin_date', 'end_date', 'duration', 'meeting_point',
                   'start_time', 'end_time', 'contact', 'email', 'website', 'organizer', 'speaker',
                   'type', 'accessibility', 'capacity', 'booking', 'target_audience',
                   'practical_info', 'approved', 'source', 'portal', 'geom', 'eid', 'structure', 'bookable',
                   'cancelled', 'cancellation_reason']
         model = TouristicEvent
+        widgets = {'geom': AutoLocateMapWidget()}
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)